+++ CHANGELOG.rst
@@ -23,6 +23,18 @@ Removed
 Fixed
 -------
 
+[0.11.2] - XXXXXX
+^^^^^^^^^^^^^^^^^
+
+Added
+-----
+- callback output channel, receives messages and uses a REST endpoint to
+  respond with messages
+
+Changed
+-------
+- channel input creation moved to the channel, every channel can now
+  customize how it gets created from the credentials file
 
 [0.11.1] - 2018-08-30
 ^^^^^^^^^^^^^^^^^^^^^